Redivide software responsibilities into 4 layers
Redesign software structure. Where the old
structure had a PathPlanner giving waypoints to
a RobotNavigator using a DriveSystem to calculate
motor powers, the new structure has a PathPlanner
giving goals to a RobotNavigator passing
waypoints to a RobotPilot which builds and
executes motions on a DriveSystem.

Add MatrixMagic to hold static methods for matrix
manipulation.
Move from arrays of values to maps from motor
names to values in MecanumDriveSystem.
Fix incorrect Mecanum wheel motor calculations
where axial motion would be negated in some
motors.
Remove MotorActionState as the premise on which
it was based, running motors at constant speeds
until precalculated encoder values are reached,
can't support linear swivels and is overly
restrictive regardless.
Change the responsibilities of DriveSystem.
Instead of being a calculator for
MotorActionStates, it is now stateful and is used
by the pilot to execute incremental motions.
Change or add OpModes to test each layer of the
new system.
Rewrite BlindPathPlanner to implement the new
responsibilities of the PathPlanner.