Add RobotHook
Add RobotHook, which encapsulates a "hook" device
that allows the robot to hang from the riggings.
Update GamepadInputInfo to implement
HookInputInfo. Currently left bumper primes or
hooks if primed and left trigger retreats to last
prime or zero state.
Rename TestMotorOpMode_Linear to
TestMotorTeleOpMode in line with the rest of the
OpModes.